renaming make.py to pack.py + update
